requirements.txt
django==5.1.1
python-dotenv==1.0.0
djangorestframework==3.15.2